Refactor grid options by replacing `MeshGridOptions` with `GridOptions`
Unified grid configuration into the `GridOptions` struct, replacing the previously used `MeshGridOptions`. Updated enum and struct implementations to improve clarity and reduce redundancy. Adjusted relevant function calls and tests to align with the new structure.
