+++ crates/burn-tensor/src/tensor/grid/mod.rs
@@ -2,16 +2,6 @@ mod meshgrid;
 
 pub use meshgrid::*;
 
-/// Enum to specify grid sparsity mode.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum GridSparsity {
-    /// The grid is sparse, expanded only at the cardinal dimensions.
-    Sparse,
-
-    /// The grid is fully expanded to the full cartesian product shape.
-    Dense,
-}
-
 /// Enum to specify index cardinal layout.
 #[derive(Default, Debug, Clone, Copy, PartialEq, Eq)]
 pub enum GridIndexing {
@@ -20,7 +10,45 @@ pub enum GridIndexing {
     #[default]
     Matrix,
 
-    /// The first two dimensions are swapped.
+    /// The same as Matrix, but the first two dimensions are swapped.
     /// Equivalent to "xy" indexing in NumPy and PyTorch.
     Cartesian,
+}
+
+/// Enum to specify grid sparsity mode.
+#[derive(Default, Debug, Clone, Copy, PartialEq, Eq)]
+pub enum GridSparsity {
+    /// The grid is fully expanded to the full cartesian product shape.
+    #[default]
+    Dense,
+
+    /// The grid is sparse, expanded only at the cardinal dimensions.
+    Sparse,
+}
+
+/// Grid policy options.
+#[derive(new, Default, Debug, Copy, Clone)]
+pub struct GridOptions {
+    /// Indexing mode.
+    pub indexing: GridIndexing,
+
+    /// Sparsity mode.
+    pub sparsity: GridSparsity,
+}
+
+impl From<GridIndexing> for GridOptions {
+    fn from(value: GridIndexing) -> Self {
+        Self {
+            indexing: value,
+            ..Default::default()
+        }
+    }
+}
+impl From<GridSparsity> for GridOptions {
+    fn from(value: GridSparsity) -> Self {
+        Self {
+            sparsity: value,
+            ..Default::default()
+        }
+    }
 }
